requirements.txt
altair==4.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cycler
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
fonttools==4.25.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver
MarkupSafe==2.1.1
matplotlib
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook-shim==0.1.0
numpy
olefile
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing
pyrsistent==0.18.1
python-dateutil
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
rich==12.6.0
semver==2.13.0
Send2Trash==1.8.0
six
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
streamlit==1.15.1
terminado==0.16.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado
tqdm
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
zipp==3.9.0
